moves table functions from utils
